mediashield/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
aiosqlite
imagehash
Pillow
numpy
transformers
torch
torchvision
chromadb
fpdf2
python-dotenv
